Removed task_repository and merged it with proxy_data_source
