+++ analysis_options.yaml
@@ -2,4 +2,5 @@ include: package:flutter_lints/flutter.yaml
 
 linter:
   rules:
-    - avoid_relative_lib_imports
+    - always_use_package_imports
+    - require_trailing_commas
