+++ lib/service_locator.dart
@@ -2,7 +2,6 @@ import 'package:device_info_plus/device_info_plus.dart';
 import 'package:dio/dio.dart';
 import 'package:get_it/get_it.dart';
 import 'package:todo_list_yandex_school_2024/data/datasources/local_data_source.dart';
-import 'package:todo_list_yandex_school_2024/data/datasources/proxy_data_source.dart';
 import 'package:todo_list_yandex_school_2024/data/datasources/remote_data_source.dart';
 import 'package:todo_list_yandex_school_2024/data/task_repository.dart';
 import 'package:uuid/uuid.dart';
@@ -14,10 +13,6 @@ void setupLocator() {
   getIt.registerLazySingleton<LocalDataSource>(() => LocalDataSource());
   getIt.registerLazySingleton<RemoteDataSource>(
       () => RemoteDataSource(getIt<Dio>()));
-  getIt.registerLazySingleton<ProxyDataSource>(() => ProxyDataSource(
-        getIt<LocalDataSource>(),
-        getIt<RemoteDataSource>(),
-      ));
       
 
   // Register repository
